Worker mining bug.
Definitions were not mutually exclusive. Care was taken to make them mutually exclusive.

Some problems related to morphing buildings remain.

+++ Source/CUNYAIModule.cpp
@@ -777,6 +777,9 @@ void CUNYAIModule::onFrame()
                             }
                             else {
                                 buildorder.clearRemainingBuildOrder(); // Neutralize the build order if something other than a worker scout is happening.
+                                if (_ANALYSIS_MODE) {
+                                    Broodwar->sendText("Clearing Build Order, board state is dangerous.");
+                                }
                             }
                         }
 
@@ -869,15 +872,37 @@ void CUNYAIModule::onFrame()
 
             bool want_gas = gas_starved && (Count_Units(UnitTypes::Zerg_Extractor, inventory) - Count_Units_In_Progress(UnitTypes::Zerg_Extractor, inventory)) > 0;  // enough gas if (many critera), incomplete extractor, or not enough gas workers for your extractors.  
             bool too_much_gas = 1 - inventory.getLn_Gas_Ratio() > delta;
-            bool no_recent_worker_alteration = miner.time_of_last_purge_ < Broodwar->getFrameCount() - 12 && miner.time_since_last_command_ > 12;
+            bool no_recent_worker_alteration = miner.time_of_last_purge_ < t_game - 12 && miner.time_since_last_command_ > 12;
+
+            // Identify old mineral task. If there's no new better job, put them back on this without disturbing them.
+            bool was_gas = miner.isAssignedGas(land_inventory);
+            bool was_mineral = miner.isAssignedMining(land_inventory);
+            bool was_long_mine = miner.isLongRangeLock(land_inventory);
+            Unit old_mineral_patch = nullptr;
+            if ((was_mineral || was_gas) && !was_long_mine) {
+                old_mineral_patch = miner.locked_mine_;
+            }
 
             if ( !isRecentCombatant(miner.bwapi_unit_) && !miner.isAssignedClearing(land_inventory) && !miner.isAssignedBuilding() && spamGuard(miner.bwapi_unit_) ) { //Do not disturb fighting workers or workers assigned to clear a position. Do not spam. Allow them to remain locked on their task. 
-                if (isEmptyWorker(u) && miner.isAssignedResource(land_inventory) && !miner.isAssignedBuilding() && my_reservation.last_builder_sent_ < t_game - Broodwar->getLatencyFrames() -  15 * 24 && !build_check_this_frame) { //only get those that are in line or gathering minerals, but not carrying them. This always irked me.
+
+                // Each mineral-related subtask does the following:
+                // Checks if it is doing a task of lower priority.
+                // It clears the worker. 
+                // It tries to assign the worker to the new task.
+                // If it is successfully assigned, continue. On the next frame you will be caught by "Maintain the locks" step.
+                // If it is not successfully assigned, return to old task.
+
+                //BUILD RELATED TASKS:
+                if (isEmptyWorker(u) && miner.isAssignedResource(land_inventory) && !miner.isAssignedBuilding() && my_reservation.last_builder_sent_ < t_game - Broodwar->getLatencyFrames() -  15 * 24 && !build_check_this_frame ) { //only get those that are in line or gathering minerals, but not carrying them. This always irked me.
                     build_check_this_frame = true;
-                    //friendly_inventory.purgeWorkerRelationsNoStop(u, land_inventory, inventory, my_reservation); //Must be disabled or else under some conditions, we "stun" a worker every frame. Usually the exact same one, essentially killing it.
+                    friendly_inventory.purgeWorkerRelationsNoStop(u, land_inventory, inventory, my_reservation); //Must be disabled or else under some conditions, we "stun" a worker every frame. Usually the exact same one, essentially killing it.
                     Building_Begin(u, inventory, enemy_inventory, friendly_inventory); // something's funny here. I would like to put it in the next line conditional but it seems to cause a crash when no major buildings are left to build.
                     if (miner.isAssignedBuilding() ) { //Don't purge the building relations here - we just established them!
                         miner.stopMine(land_inventory);
+                        continue;
+                    }
+                    else if (old_mineral_patch){
+                        attachToParticularMine(old_mineral_patch, land_inventory, miner); // go back to your old job. Updated unit.
                         continue;
                     }
                 } // Close Build loop
@@ -892,6 +917,9 @@ void CUNYAIModule::onFrame()
                     continue;
                 }
 
+
+
+               //MINERAL RELATED TASKS
                //Workers need to clear empty patches.
                 bool time_to_start_clearing_a_path = inventory.hatches_ >= 2 && Nearby_Blocking_Minerals(u, friendly_inventory);
                 if (time_to_start_clearing_a_path && inventory.workers_clearing_ == 0 && isEmptyWorker(u)) {
@@ -901,20 +929,18 @@ void CUNYAIModule::onFrame()
                         inventory.updateWorkersClearing(friendly_inventory, land_inventory);
                         continue;
                     }
+                    else if (old_mineral_patch) {
+                        attachToParticularMine(old_mineral_patch, land_inventory, miner); // go back to your old job. Updated unit.
+                        continue;
+                    }
                 } // clear those empty mineral patches that block paths.
 
                 // Gives all loose workers a target mine. Maintains gas/mineral balance. 
                 land_inventory.countViableMines();
                 bool could_use_another_gas = land_inventory.local_gas_collectors_ * 2 <= land_inventory.local_refineries_;
-                bool worker_might_be_in_bad_task = (want_gas && miner.isAssignedMining(land_inventory) && could_use_another_gas) || ((!want_gas || too_much_gas) && miner.isAssignedGas(land_inventory));
+                bool worker_might_be_in_bad_task = (want_gas && miner.isAssignedMining(land_inventory) && could_use_another_gas) || ((!want_gas || too_much_gas) && miner.isAssignedGas(land_inventory)) || was_long_mine;
 
                 if ( (!miner.isAssignedResource(land_inventory) && !miner.isAssignedBuilding()) || (( worker_might_be_in_bad_task && inventory.last_gas_check_ < t_game - 5 * 24) && isEmptyWorker(u)) ) { //if this is your first worker of the frame consider resetting him.
-                    bool was_gas = miner.isAssignedGas(land_inventory);
-                    bool was_mineral = miner.isAssignedResource(land_inventory);
-                    Unit old_mineral_patch = nullptr;
-                    if (was_mineral) {
-                        old_mineral_patch = miner.locked_mine_;
-                    }
 
                     friendly_inventory.purgeWorkerRelationsNoStop(u, land_inventory, inventory, my_reservation);
                     inventory.last_gas_check_ = t_game;
@@ -924,8 +950,7 @@ void CUNYAIModule::onFrame()
                             continue;
                         }
                         else if (old_mineral_patch) {
-                            miner.startMine(land_inventory.resource_inventory_.find(old_mineral_patch)->second, land_inventory); // go back to your old job.
-                            miner.updateStoredUnit(u);
+                            attachToParticularMine(old_mineral_patch, land_inventory, miner); // go back to your old job. Updated unit.
                             continue;
                         }
                         else { // default to gathering minerals.
@@ -935,9 +960,8 @@ void CUNYAIModule::onFrame()
                             }
                         }
                     }
-                    else if (old_mineral_patch) {
-                        miner.startMine(land_inventory.resource_inventory_.find(old_mineral_patch)->second, land_inventory); // go back to your old job.
-                        miner.updateStoredUnit(u);
+                    else if (old_mineral_patch && !was_gas) { // If gas was something to avoid, don't go back onto it.
+                        attachToParticularMine(old_mineral_patch, land_inventory, miner); // go back to your old job. Updated unit.
                         continue;
                     }
                     else { // default to gathering minerals.
@@ -953,7 +977,12 @@ void CUNYAIModule::onFrame()
             if (!isEmptyWorker(u) && u->isIdle() && no_recent_worker_alteration) {
                 friendly_inventory.purgeWorkerRelationsNoStop(u, land_inventory, inventory, my_reservation); //If he can't get back to work something's wrong with you and we're resetting you.
                 miner.bwapi_unit_->returnCargo();
-                miner.updateStoredUnit(u);
+                if (old_mineral_patch) {
+                    attachToParticularMine(old_mineral_patch, land_inventory, miner); // go back to your old job. Updated unit.
+                }
+                else {
+                    miner.updateStoredUnit(u);
+                }
                 continue;
             }
 
@@ -961,6 +990,9 @@ void CUNYAIModule::onFrame()
             if ( u->isIdle() && no_recent_worker_alteration) {
                 friendly_inventory.purgeWorkerRelationsNoStop(u, land_inventory, inventory, my_reservation);
                 Worker_Gather(u, UnitTypes::Resource_Mineral_Field, friendly_inventory);
+                if (!miner.isAssignedMining(land_inventory)) {
+                    Worker_Gather(u, UnitTypes::Resource_Vespene_Geyser, friendly_inventory);
+                }
                 miner.updateStoredUnit(u);
             }
 
@@ -994,13 +1026,13 @@ void CUNYAIModule::onFrame()
                 miner.updateStoredUnit(u);
                 continue;
             } 
-            else if (worker_has_lockable_task && miner.isMovingLock(land_inventory)) {
-                if (!miner.bwapi_unit_->gather(miner.locked_mine_)) { // reassign him back to work.
-                    friendly_inventory.purgeWorkerRelationsNoStop(u, land_inventory, inventory, my_reservation); //If he can't get back to work something's wrong with you and we're resetting you.
-                }
-                miner.updateStoredUnit(u);
-                continue;
-            }
+            //else if (worker_has_lockable_task && miner.isMovingLock(land_inventory)) {
+            //    if (!miner.bwapi_unit_->gather(miner.locked_mine_)) { // reassign him back to work.
+            //        friendly_inventory.purgeWorkerRelationsNoStop(u, land_inventory, inventory, my_reservation); //If he can't get back to work something's wrong with you and we're resetting you.
+            //    }
+            //    miner.updateStoredUnit(u);
+            //    continue;
+            //}
 
         } // Close Worker management loop
         auto end_worker = std::chrono::high_resolution_clock::now();
@@ -1010,11 +1042,13 @@ void CUNYAIModule::onFrame()
 
         bool unconsidered_unit_type = std::find(types_of_units_checked_for_upgrades_this_frame.begin(), types_of_units_checked_for_upgrades_this_frame.end(), u_type) == types_of_units_checked_for_upgrades_this_frame.end();
 
-        if ( isIdleEmpty( u ) && !u->canAttack() && u_type != UnitTypes::Zerg_Larva && unconsidered_unit_type && // no trying to morph hydras anymore.
+        if ( isIdleEmpty( u ) && !u->canAttack() && u_type != UnitTypes::Zerg_Larva && unconsidered_unit_type && spamGuard(u) &&
             (u->canUpgrade() || u->canResearch() || u->canMorph()) ) { // this will need to be revaluated once I buy units that cost gas.
 
             if (Tech_Begin(u, friendly_inventory, inventory)) {
-                friendly_inventory.purgeWorkerRelationsNoStop(u, land_inventory, inventory, my_reservation);
+                if (u_type.isWorker()) {
+                    friendly_inventory.purgeWorkerRelationsNoStop(u, land_inventory, inventory, my_reservation);
+                }
                 types_of_units_checked_for_upgrades_this_frame.push_back(u_type); // only check each type once.
             }
             //PrintError_Unit( u );
@@ -1024,7 +1058,7 @@ void CUNYAIModule::onFrame()
         //Creep Colony upgrade loop.  We are more willing to upgrade them than to build them, since the units themselves are useless in the base state.
         auto start_creepcolony = std::chrono::high_resolution_clock::now();
 
-        if ( u_type == UnitTypes::Zerg_Creep_Colony ) {
+        if ( u_type == UnitTypes::Zerg_Creep_Colony && spamGuard(u) ) {
             if (u->getDistance(mutating_creep_colony_position) < UnitTypes::Zerg_Sunken_Colony.sightRange() && mutating_creep_colony_type == UnitTypes::Zerg_Sunken_Colony ) {
                 Check_N_Build(UnitTypes::Zerg_Sunken_Colony, u, friendly_inventory, true);
                 mutating_creep_colony_position = u->getPosition();
@@ -1040,8 +1074,8 @@ void CUNYAIModule::onFrame()
             else if (!mutating_creep_this_frame){
                 Unit_Inventory local_e = getUnitInventoryInRadius(enemy_inventory, u->getPosition(), inventory.my_portion_of_the_map_);
                 local_e.updateUnitInventorySummary();
-                bool can_sunken = Count_Units(UnitTypes::Zerg_Spawning_Pool, friendly_inventory) > 0;
-                bool can_spore = Count_Units(UnitTypes::Zerg_Evolution_Chamber, friendly_inventory) > 0;
+                bool can_sunken = Count_Units(UnitTypes::Zerg_Spawning_Pool, inventory) > 0;
+                bool can_spore = Count_Units(UnitTypes::Zerg_Evolution_Chamber, inventory) > 0;
                 bool need_static_d = buildorder.checkBuilding_Desired(UnitTypes::Zerg_Spore_Colony) || buildorder.checkBuilding_Desired(UnitTypes::Zerg_Sunken_Colony);
                 bool want_static_d = (army_starved || local_e.stock_fighting_total_ > 0) && (can_sunken || can_spore);
                 bool u_relatively_weak_against_air = checkWeakAgainstAir(friendly_inventory, enemy_inventory) && local_e.stock_fliers_ > 0; // div by zero concern. Derivative of the above equation 
@@ -1050,8 +1084,6 @@ void CUNYAIModule::onFrame()
                     bool cloak_nearby = local_e.cloaker_count_ > 0;
                     bool local_air_problem = local_e.stock_fliers_ > 0;
                     bool global_air_problem = u_relatively_weak_against_air;
-                    buildorder.checkBuilding_Desired(UnitTypes::Zerg_Sunken_Colony);
-                    buildorder.checkBuilding_Desired(UnitTypes::Zerg_Spore_Colony);
                     if (can_sunken && can_spore) {
                         if (local_air_problem || global_air_problem || cloak_nearby) { // if they have a flyer (that can attack), get spores.
                             Check_N_Build(UnitTypes::Zerg_Spore_Colony, u, friendly_inventory, true);
@@ -1393,11 +1425,13 @@ void CUNYAIModule::onUnitMorph( BWAPI::Unit unit )
         }
     }
 
-    if ( unit->getType().isWorker() || unit->getBuildType().isBuilding() ) {
+    if ( unit->getType().isWorker() ) {
         friendly_inventory.purgeWorkerRelations(unit, land_inventory, inventory, my_reservation);
     }
 
+
     if ( unit->getBuildType().isBuilding() ) {
+        friendly_inventory.purgeWorkerRelationsNoStop(unit, land_inventory, inventory, my_reservation);
         buildorder.updateRemainingBuildOrder(unit->getBuildType());
     }
